bot/requirements.txt
asyncio
discord
DiscordUtils
PyNaCl
python-dotenv
python-socketio[asyncio_client]
requests
yt-dlp